switch to no separate compileutils file: faster
